Add emoji support

+++ bot.py
@@ -27,25 +27,32 @@ def start(bot, update):
     update.message.reply_text('Olá! Use o comando /quote em resposta a uma mensagem')
 
 def get_user_pic(user: User):
+    name = user.first_name + " " + user.last_name if user.last_name else user.first_name
+    name = name.split(" ")
+    formated_name = name[1].upper() + ", " + name[0] if len(name) > 1 else name[0]
+
     photos = user.get_profile_photos().photos
 
     if not photos or not photos[0]:
-        return None
+        return formated_name
 
     # first photo, last version (better quality)
     photo_size = photos[0][-1]
     photo = photo_size.get_file().download_as_bytearray()
-
-    name = user.first_name + " " + user.last_name if user.last_name else user.first_name
-    name = name.split(" ")
-    formated_name = name[1].upper() + ", " + name[0] if len(name) > 1 else name[0]
 
     return (photo, (photo_size.width, photo_size.height), formated_name)
 
 
 def apply_overlay(n, photo, quote, name):
     result = f'result-{n}.jpg'
-    command = ['bash', 'make_image.sh', quote, name, photo, result]
+
+    if photo:
+        script = 'make_image.sh'
+    else:
+        script = 'make_image_noprofile.sh'
+        photo = 'None'
+
+    command = ['bash', script, quote, name, photo, result]
     subprocess.call(command)
 
     return result
@@ -57,7 +64,7 @@ def make_quote(bot, update):
     if update['message']['forward_from'] is not None:
         if update['message']['chat']['type'] == 'group':
             return
-        pic, size, name = get_user_pic(update.message.forward_from)
+        user_pic = get_user_pic(update.message.forward_from)
         quote = update['message']['text']
     else:
         if update['message']['reply_to_message'] is None:
@@ -65,20 +72,27 @@ def make_quote(bot, update):
             return
 
         if update['message']['reply_to_message']['forward_from'] is not None:
-            pic, size, name = get_user_pic(update.message.reply_to_message.forward_from)
+            user_pic = get_user_pic(update.message.reply_to_message.forward_from)
         else:
-            pic, size, name = get_user_pic(update.message.reply_to_message.from_user)
+            user_pic = get_user_pic(update.message.reply_to_message.from_user)
         quote = update['message']['reply_to_message']['text']
 
     n = random.randint(1, 1000)
-    file_name = f'user-{n}.jpg'
-    open(file_name, 'wb').write(pic)
+    file_name = False
+
+    if len(user_pic) == 3:
+        file_name = f'user-{n}.jpg'
+        open(file_name, 'wb').write(user_pic[0])
+        name = user_pic[2]
+    else:
+        name = user_pic
 
     result = apply_overlay(n, file_name, quote, name)
 
     update.message.reply_photo(photo=open(result, 'rb'))
 
-    os.remove(file_name)
+    if file_name:
+        os.remove(file_name)
     os.remove(result)
